+++ Applications/StepSync/desc.sh
@@ -1,13 +1,14 @@
 #!/bin/sh
 
-FWNAME="RSSKit"
-VERSION="0.6"
-SHORTNAME=""
+APPNAME="StepSync"
+VERSION="0.1"
+SHORTNAME="stepsync"
 REPO="gap"
 METHOD="git"					# "git" | "get"
 HUB="https://github.com/gnustep/"
 EXT=""					# if get method: "zip" | "tar.gz"...
-PATH_SRC="libs/${FWNAME}"
+PATH_SRC="system-apps/${APPNAME}"
+FRAMEWORKS=""
 DEPS=""					# "YES" or leave empty
 
 # READ SPECIFIC INFOS IN SOURCE FOLDER
@@ -15,3 +16,4 @@ BUILD_ARGS=""
 INSTALL_ARGS="${BUILD_ARGS}"
 
 PATCH=""			# path of the file.patch or leave empty
+# RUN_AT_END="YES"				# "YES" | comment with #